refactor: 튜토리얼 모달 함수 분리
